Changed pixels in Frame from int to uint8_t

+++ src/frame/Frame.cpp
@@ -15,7 +15,7 @@ int Frame::xyToIndexZ(int x, int y) const {
 Frame::Frame(int width, int height) {
     this->width = width;
     this->height = height;
-    pixels = new int[xyToIndex(width, height)];
+    pixels = new uint8_t[xyToIndex(width, height)];
     zBuffer = new float[xyToIndexZ(width, height)];
 }
